CORRIGIDO REGISTRO

+++ src/main/java/fiap/com/br/gestaoresiduos/repository/EndercoRepository.java
@@ -3,5 +3,5 @@ package fiap.com.br.gestaoresiduos.repository;
 import fiap.com.br.gestaoresiduos.model.Endereco;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EndercoRepository extends JpaRepository<Endereco, Integer> {
+public interface EndercoRepository extends JpaRepository<Endereco, Long> {
 }
